requirements/base.txt
Django==1.8.18
djangorestframework==3.6.3
mysqlclient==1.3.10
